renamed the SVG related functions so that they deal with an anonymous data
2004-06-30  Sven Neumann  <sven@gimp.org>

	* app/widgets/gimpselectiondata.[ch]: renamed the SVG related
	functions so that they deal with an anonymous data stream that
	could as well be a PNG image.

	* app/widgets/gimpdnd.[ch]
	* app/widgets/gimpcontainertreeview-dnd.c: changed accordingly.

	* app/display/gimpdisplayshell-dnd.[ch]
	* app/vectors/gimpvectors-import.[ch]
	* app/widgets/gimpcontainertreeview-dnd.c
	* app/widgets/gimpvectorstreeview.c: use gsize for the length of
	the buffer.

	* app/widgets/gimpdnd.[ch]
	* app/widgets/widgets-enums.[ch]: added GIMP_DND_TYPE_PNG which isn't
	used yet.

+++ app/display/gimpdisplayshell-dnd.c
@@ -152,7 +152,7 @@ gimp_display_shell_drop_vectors (GtkWidget    *widget,
 void
 gimp_display_shell_drop_svg (GtkWidget     *widget,
                              const gchar   *svg_data,
-                             gint           svg_data_len,
+                             gsize          svg_data_len,
                              gpointer       data)
 {
   GimpDisplayShell *shell  = GIMP_DISPLAY_SHELL (data);
