+++ app/display/gimpdisplayshell-dnd.h
@@ -28,7 +28,7 @@ void   gimp_display_shell_drop_vectors  (GtkWidget     *widget,
                                          gpointer       data);
 void   gimp_display_shell_drop_svg      (GtkWidget     *widget,
                                          const gchar   *svg_data,
-                                         gint           svg_data_length,
+                                         gsize          svg_data_length,
                                          gpointer       data);
 void   gimp_display_shell_drop_pattern  (GtkWidget     *widget,
                                          GimpViewable  *viewable,
